refactor userBalances script into modules
